move DQM/SiPixelPhase1Analyzer to DQM/TrackerRemapper
