--- DQM/SiPixelPhase1Analyzer/python/CfiFile_cfi.py
@@ -1,4 +1,0 @@
-import FWCore.ParameterSet.Config as cms
-
-demo = cms.EDAnalyzer('SiPixelPhase1Analyzer'
-)
