+++ DQM/TrackerRemapper/plugins/BuildFile.xml
@@ -1,14 +1,14 @@
-<use name="FWCore/Framework"/>
-<use name="FWCore/PluginManager"/>
-<use name="FWCore/ParameterSet"/>
-<use name="Geometry/Records"/> 
-<use name="Geometry/TrackerGeometryBuilder"/>
-<use name="Geometry/CommonTopologies"/>
+<use name="CalibTracker/SiStripCommon"/>
 <use name="CommonTools/UtilAlgos"/>
 <use name="DataFormats/GeometrySurface"/>
-<use name="RecoTracker/TkDetLayers"/>
+<use name="DataFormats/TrackReco"/>
 <use name="DataFormats/TrackerCommon"/>
 <use name="DataFormats/TrackerRecHit2D"/>
-<use name="DataFormats/TrackReco"/>
-<use name="CalibTracker/SiStripCommon"/>
+<use name="FWCore/Framework"/>
+<use name="FWCore/ParameterSet"/>
+<use name="FWCore/PluginManager"/>
+<use name="Geometry/CommonTopologies"/>
+<use name="Geometry/Records"/> 
+<use name="Geometry/TrackerGeometryBuilder"/>
+<use name="RecoTracker/TkDetLayers"/>
 <flags EDM_PLUGIN="1"/>
